build: Move away from deprecated path
