+++ demo/GoDemoApp.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/google/uuid"
-	"github.com/open-feature/go-sdk/pkg/openfeature"
+	"github.com/open-feature/go-sdk/openfeature"
 	confidence "github.com/spotify/confidence-openfeature-provider-go/pkg/provider"
 )
